Issue #39 : Metoda za ponistavanje dogadjaja

+++ Aplikacija/Softver_za_prodaju_ulaznica/src/users/RegularUser.java
@@ -1,17 +1,17 @@
 package users;
 
 public class RegularUser extends User {
-	private int credits = 1000;
+	private double credits = 1000;
 	
 	public RegularUser(String name, String surname, String username, String password) {
 		super(name,surname,username,password);
 	}
 	
-	public void addCredits(int credits) {
+	public void addCredits(double credits) {
 		this.credits += credits;
 	}
 	
-	public boolean takeCredits(int credits) {
+	public boolean takeCredits(double credits) {
 		if(this.credits < credits) {
 			return false;
 		}
